Split UI components into modules
Move the monolithic src/ui/components.rs into focused modules:
button, card, input, scrollable, sensors, text and a new mod.rs.
Introduce ButtonVisual and primary_button APIs, and adapt call
sites (UiTheme/Icon imports and component usages) across the UI.
