+++ src/ui/settings/sidebar.rs
@@ -4,25 +4,18 @@ use iced::{Background, Border, Color, Element, Length};
 
 use crate::app::{Message, Snapdash};
 use crate::theme::{Palette, metric};
-use crate::ui::components;
 
 use super::SettingsPage;
 
 pub fn view<'a>(snap: &'a Snapdash) -> Element<'a, Message> {
     let p = snap.theme.palette();
 
-    let search: Element<Message> =
-        components::mac_input("Search settings ...", &snap.settings_search, p)
-            .on_input(Message::SettingsSearchChanged)
-            .width(Length::Fill)
-            .into();
-
     let mut nav = column![].spacing(2);
     for &page in SettingsPage::ALL {
         nav = nav.push(nav_item(page, snap.settings_page == page, p));
     }
 
-    let inner = column![search, nav].spacing(metric::GAP);
+    let inner = column![nav].spacing(metric::GAP);
 
     iced::widget::container(inner)
         .width(200)
